# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArtifactFlow API",
    "description": "Multi-agent system API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/tools/units": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Units",
        "operationId": "list_units_api_v1_admin_tools_units_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Unit",
        "operationId": "create_unit_api_v1_admin_tools_units_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateToolUnitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/tools/units/import": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Import Unit Seed",
        "operationId": "import_unit_seed_api_v1_admin_tools_units_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_unit_seed_api_v1_admin_tools_units_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/tools/units/{name}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Unit",
        "operationId": "get_unit_api_v1_admin_tools_units__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Unit",
        "operationId": "update_unit_api_v1_admin_tools_units__name__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToolUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Unit",
        "operationId": "delete_unit_api_v1_admin_tools_units__name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tools/units/{name}/export": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Export Unit Seed",
        "operationId": "export_unit_seed_api_v1_admin_tools_units__name__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tools/units/{name}/test": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Test Unit Connection",
        "operationId": "test_unit_connection_api_v1_admin_tools_units__name__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolUnitTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tools/agents": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_api_v1_admin_tools_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/tools/units/{name}/agents/{agent_name}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Mount Unit",
        "operationId": "mount_unit_api_v1_admin_tools_units__name__agents__agent_name__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MountUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Unmount Unit",
        "operationId": "unmount_unit_api_v1_admin_tools_units__name__agents__agent_name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tools/units/{name}/credentials/{placeholder}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Set Credential",
        "operationId": "set_credential_api_v1_admin_tools_units__name__credentials__placeholder__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "placeholder",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "title": "Placeholder"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Credential",
        "operationId": "delete_credential_api_v1_admin_tools_units__name__credentials__placeholder__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "placeholder",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "title": "Placeholder"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentListResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSummaryResponse"
            },
            "type": "array",
            "title": "Agents"
          }
        },
        "type": "object",
        "required": [
          "agents"
        ],
        "title": "AgentListResponse"
      },
...
      "Body_import_unit_seed_api_v1_admin_tools_units_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_unit_seed_api_v1_admin_tools_units_import_post"
      },
...
      "CreateToolUnitRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "title": "Name",
            "description": "unit \u540d,\u5168\u5c40\u552f\u4e00,\u7981\u542b '__'"
          },
          "kind": {
            "type": "string",
            "enum": [
              "tool",
              "toolset",
              "mcp"
            ],
            "title": "Kind",
            "default": "tool"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "department"
            ],
            "title": "Visibility",
            "default": "public"
          },
          "defer": {
            "type": "boolean",
            "title": "Defer",
            "default": false
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/ToolMemberSpec"
            },
            "type": "array",
            "title": "Members"
          },
          "provider_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpProviderConfigSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateToolUnitRequest",
        "description": "POST /api/v1/admin/tools/units \u2014\u2014 \u65b0\u5efa dynamic unit\u3002"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "required": [
          "detail"
        ]
      },
...
      "MountResponse": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "unit_name": {
            "type": "string",
            "title": "Unit Name"
          },
          "member_state": {
            "type": "string",
            "title": "Member State"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "agent_name",
          "unit_name",
          "member_state",
          "source"
        ],
        "title": "MountResponse",
        "description": "PUT .../agents/{agent} \u7684\u8fd4\u56de(\u6302\u8f7d/\u6539\u6210\u5458\u6001\u540e\u7684\u7ed1\u5b9a\u5feb\u7167)\u3002"
      },
      "MountUnitRequest": {
        "properties": {
          "member_state": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Member State",
            "default": "enabled"
          }
        },
        "type": "object",
        "title": "MountUnitRequest",
        "description": "PUT /api/v1/admin/tools/units/{name}/agents/{agent_name}"
      },
      "MountedAgentResponse": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "member_state": {
            "type": "string",
            "title": "Member State"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "agent_name",
          "member_state",
          "source"
        ],
        "title": "MountedAgentResponse"
      },
...
      "SetCredentialRequest": {
        "properties": {
          "value": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Value",
            "description": "\u660e\u6587,\u52a0\u5bc6\u843d\u5e93\u540e\u6c38\u4e0d\u56de\u8bfb"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "SetCredentialRequest",
        "description": "PUT /api/v1/admin/tools/units/{name}/credentials/{placeholder} \u2014\u2014 \u5199-only\u3002"
      },
...
      "ToolMemberSpec": {
        "properties": {
          "member_name": {
            "type": "string",
            "maxLength": 64,
            "title": "Member Name",
            "description": "\u4f5c\u8005\u88f8\u540d;singleton \u4f1a\u88ab\u89c4\u6574\u4e3a unit \u540d"
          },
          "permission": {
            "type": "string",
            "enum": [
              "auto",
              "confirm"
            ],
            "title": "Permission",
            "default": "confirm"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "default": ""
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "GET"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "response_extract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Extract"
          },
          "artifact_output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactOutputSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeout": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Timeout",
            "default": 60
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "member_name",
          "permission",
          "description",
          "endpoint",
          "method",
          "headers",
          "input_schema",
          "response_extract",
          "artifact_output",
          "timeout"
        ],
        "title": "ToolMemberSpec"
      },
      "ToolUnitImportResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "imported",
            "title": "Status",
            "default": "imported"
          },
          "unit": {
            "$ref": "#/components/schemas/ToolUnitResponse"
          }
        },
        "type": "object",
        "required": [
          "status",
          "unit"
        ],
        "title": "ToolUnitImportResponse"
      },
      "ToolUnitListResponse": {
        "properties": {
          "units": {
            "items": {
              "$ref": "#/components/schemas/ToolUnitResponse"
            },
            "type": "array",
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "units"
        ],
        "title": "ToolUnitListResponse"
      },
      "ToolUnitResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "defer": {
            "type": "boolean",
            "title": "Defer"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "provider_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Config"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/ToolMemberResponse"
            },
            "type": "array",
            "title": "Members"
          },
          "mounted_agents": {
            "items": {
              "$ref": "#/components/schemas/MountedAgentResponse"
            },
            "type": "array",
            "title": "Mounted Agents"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/CredentialStatusResponse"
            },
            "type": "array",
            "title": "Credentials"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "description",
          "visibility",
          "defer",
          "provider",
          "provider_config",
          "source",
          "members",
          "mounted_agents",
          "credentials"
        ],
        "title": "ToolUnitResponse"
      },
      "ToolUnitTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "tool_count": {
            "type": "integer",
            "title": "Tool Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "tool_count"
        ],
        "title": "ToolUnitTestResponse"
      },
...
      "UpdateToolUnitRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "title": "Name",
            "description": "unit \u540d,\u5168\u5c40\u552f\u4e00,\u7981\u542b '__'"
          },
          "kind": {
            "type": "string",
            "enum": [
              "tool",
              "toolset",
              "mcp"
            ],
            "title": "Kind",
            "default": "tool"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "department"
            ],
            "title": "Visibility",
            "default": "public"
          },
          "defer": {
            "type": "boolean",
            "title": "Defer",
            "default": false
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/ToolMemberSpec"
            },
            "type": "array",
            "title": "Members"
          },
          "provider_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpProviderConfigSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateToolUnitRequest",
        "description": "PUT /api/v1/admin/tools/units/{name} \u2014\u2014 \u6574\u4f53\u66ff\u6362 dynamic unit(name \u53d6\u81ea\u8def\u5f84)\u3002"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}